// src/modules/force-join/enums/force-join-callback-data.enum.ts
export enum ForceJoinCallbackData {
  AddChannel = 'add_force_join_cannel',
  GetChannels = 'get_force_join_channels',
  ShowChannelDetails = 'show_force_join_channel_details:',
  ActiveChannel = 'active_force_join_channel:',
  InactiveChannel = 'inactive_force_join_channel:',
  RemoveChannel = 'remove_force_join_channel:',
  AddToList = 'add_channel_to_force_join:',
  ShowTimeRemaining = 'show_time_remaining_force_join_channel:',
}